Adds config test to ensure docker is installed
We'll need to install docker for testing via an external role,
but otherwise the test suite is meaningless (since the role will fail
during provisioning when it tries to start a container).

Updates all apt packages

Under Jessie, the version of pip is old, and throws errors on the second
converge run for pip docker-py state=latest. Might have to add a
dependency role for updating pip.